refactoring for cruds, removal of unnecessary functionality, test coverage, linter fixes

+++ backend/app/app/models/balance.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Column, ForeignKey, DateTime, Integer
-from sqlalchemy.orm import relationship, Mapped
+from sqlalchemy import Column, ForeignKey, DateTime, Float
+from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.sql import func
 from uuid import uuid4
@@ -16,9 +16,8 @@ if TYPE_CHECKING:
 class Balance(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, index=True, default=uuid4)
     user_id = Column(UUID(as_uuid=True), ForeignKey("user.id"))
-    amount = Column(Integer, nullable=True)
-    amount_reserved = Column(Integer, nullable=True, default=0)
+    amount = Column(Float, nullable=True)
+    amount_reserved = Column(Float, nullable=True, default=0)
     last_update = Column(DateTime(timezone=True), server_default=func.now(), server_onupdate=func.now(), nullable=True)
     user = relationship("User", back_populates="balance")
     transaction_balance = relationship("TransactionBalance", back_populates='balance')
-
