+++ backend/app/app/models/transaction_event.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Column, ForeignKey, DateTime, Integer, String
+from sqlalchemy import Column, ForeignKey, DateTime, Float, String
 from sqlalchemy.orm import relationship, Mapped
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.sql import func
@@ -19,10 +19,11 @@ class TransactionEvent(Base):
 
     id = Column(UUID(as_uuid=True), primary_key=True, index=True, default=uuid4)
     transaction_id = Column(UUID(as_uuid=True), ForeignKey("transaction.id"), nullable=False)
-    type = Column(String, nullable=False) # authorization, capture, refund
-    amount = Column(Integer, nullable=False)
+    type = Column(String, nullable=False)  # authorization, capture, refund
+    amount = Column(Float, nullable=False)
     currency = Column(String(3), nullable=False, default='GBP')
-    category = Column(String, nullable=False)  # payin, payout, internal_in (receive to balance), internal_out (withdraw from balance)
+    # Categories: payin, payout, internal_in (receive to balance), internal_out (withdraw from balance)
+    category = Column(String, nullable=False)
     user_id = Column(UUID(as_uuid=True), ForeignKey("user.id"), nullable=False)
     gateway_id = Column(String)
     created = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
